3_2 แก้ int เป็น long int

+++ 6320500620_3_2.c
@@ -4,17 +4,17 @@ void main()
 {
     int n=0,i,k;
     scanf("%d",&n);
-    int p,a,getinfo[2];
+    long int p,a,getinfo[2];
     int time[48]={0};
     for(i=0 ; i<n ; i++)
     {
         for(k=0; k<2 ; k++)
-            scanf("%d",&getinfo[k]);
+            scanf("%li",&getinfo[k]);
         p=getinfo[0];
         a=getinfo[1];
         time[p-1]=time[p-1]+a;
     }
-    int max=time[0],ans=0;
+    long int max=time[0],ans=0;
     for(i=0 ; i<n ; i++)
     {
         if(time[i]>max)
@@ -23,6 +23,6 @@ void main()
             ans=i;
         }
     }
-    printf("%d %d",ans+1,max);
+    printf("%li %li",ans+1,max);
 
 }
